+++ src/main/java/frc/robot/subsystems/AutoAlign/AutoAlign.java
@@ -1,225 +1,113 @@
 package frc.robot.Subsystems.AutoAlign;
 
-import static edu.wpi.first.units.Units.*;
-import static frc.robot.GlobalConstants.ROBOT_MODE;
-import static frc.robot.Subsystems.AutoAlign.AutoAlignConstants.*;
+import java.util.function.Supplier;
 
-import choreo.trajectory.SwerveSample;
-import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.math.controller.PIDController;
-import edu.wpi.first.math.controller.ProfiledPIDController;
-import edu.wpi.first.math.filter.Debouncer;
-import edu.wpi.first.math.filter.Debouncer.DebounceType;
-import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.geometry.Translation2d;
-import edu.wpi.first.math.kinematics.ChassisSpeeds;
-import edu.wpi.first.wpilibj.DriverStation;
-import edu.wpi.first.wpilibj.DriverStation.Alliance;
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.GlobalConstants.RobotMode;
-import frc.robot.Subsystems.Drive.Drive;
-import java.util.ArrayList;
 import org.littletonrobotics.junction.Logger;
-import org.team7525.autoAlign.RepulsorFieldPlanner;
 import org.team7525.subsystem.Subsystem;
 
+import edu.wpi.first.math.MathUtil;
+import edu.wpi.first.math.controller.ProfiledPIDController;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.geometry.Transform2d;
+import edu.wpi.first.math.geometry.Translation2d;
+import edu.wpi.first.math.kinematics.ChassisSpeeds;
+import edu.wpi.first.math.trajectory.TrapezoidProfile;
+import frc.robot.Subsystems.Drive.Drive;
+
 public class AutoAlign extends Subsystem<AutoAlignStates> {
+    // 254 code (its actually so broken omg I loveeeee 254)
+    
+    public static AutoAlign instance;
 
-	private static AutoAlign instance;
+    private final ProfiledPIDController rotationController = new ProfiledPIDController(6, 0, 0, new TrapezoidProfile.Constraints(0, 0), 0.02);
+    private final ProfiledPIDController translationalController = new ProfiledPIDController(4, 0, 0, new TrapezoidProfile.Constraints(0, 0), 0.02);
 
-	private final Drive drive = Drive.getInstance();
-	private final RepulsorFieldPlanner repulsor = new RepulsorFieldPlanner(new ArrayList<>(), new ArrayList<>(), (ROBOT_MODE == RobotMode.SIM));
+    private final Drive drive = Drive.getInstance();
 
-	private PIDController translationXController;
-	private PIDController translationYController;
+    private Translation2d lastSetpointTranslation;
 
-	private ProfiledPIDController rotationController;
+    private double driveErrorAbs;
+    private double thetaErrorAbs;
 
-	private PIDController repulsionTranslationController;
-	private PIDController repulsionRotationController;
+    // 1 Meter away and more u use max ff, 0.2 meters away or less u rely on pure PID and scale down on the interval :boiled:
+    private double ffMinRadius = 0.2, ffMaxRadius = 1.0;
+    
+    private final Supplier<Pose2d> targetPoseSupplier = getState()::getTargetPose;
+    private final Supplier<Pose2d> currentPoseSupplier = drive::getPose;
 
-	private Debouncer autoAlignDebouncer;
+    
+    private AutoAlign() {
+        super("AutoAlign", AutoAlignStates.OFF);
 
-	private Pose2d targetPose;
-	private Pose2d goalPose;
-	private Pose2d reefPose = REEF_POSE;
-	private Pose2d interpolatedPose;
-	private boolean isRedAlliance = DriverStation.getAlliance().isPresent() && DriverStation.getAlliance().get() == Alliance.Red;
-	private double interpolatedDistanceFromReef;
+        this.lastSetpointTranslation = new Translation2d();
+        this.rotationController.enableContinuousInput(-Math.PI, Math.PI);
+    }
 
-	private	boolean enteredBrainDead;
-	private boolean repulsorActivated;
+    public static AutoAlign getInstance() {
+        if (instance == null) {
+            instance = new AutoAlign();
+        }
+        return instance;
+    }
 
-	private double xApplied = 0;
-	private double yApplied = 0;
-
-	private double timer;
-	private final SendableChooser<Boolean> negativeChooser = new SendableChooser<>();
-
-	private AutoAlign() {
-		super("AutoAlign", AutoAlignStates.OFF);
-		// PID Config
-		this.translationXController = TRANSLATIONALX_CONTROLLER.get();
-		this.translationYController = TRANSLATIONALY_CONTROLLER.get();
-		this.rotationController = ROTATIONAL_CONTROLLER.get();
-		this.repulsionTranslationController = REPULSOR_TRANSLATIONAL_CONTROLLER.get();
-		this.repulsionRotationController = REPULSOR_ROTATIONAL_CONTROLLER.get();
-
-		this.translationXController.setTolerance(TRANSLATIONAL_COMPONENT_ERROR_MARGIN.in(Meters));
-		this.translationYController.setTolerance(TRANSLATIONAL_COMPONENT_ERROR_MARGIN.in(Meters));
-		this.repulsionTranslationController.setTolerance(TRANSLATIONAL_COMPONENT_ERROR_MARGIN.in(Meters));
-		this.repulsionRotationController.setTolerance(ANGLE_ERROR_MARGIN.in(Radians));
-		this.rotationController.setTolerance(ANGLE_ERROR_MARGIN.in(Radians));
-
-		rotationController.enableContinuousInput(MIN_HEADING_ANGLE.in(Radian), MAX_HEADING_ANGLE.in(Radian));
-		repulsionRotationController.enableContinuousInput(MIN_HEADING_ANGLE.in(Radians), MAX_HEADING_ANGLE.in(Radians));
-
-		autoAlignDebouncer = new Debouncer(0.5, DebounceType.kRising);
-
-		repulsorActivated = false;
-		enteredBrainDead = false;
-		targetPose = new Pose2d();
-		goalPose = new Pose2d();
-
-		timer = -1;
-		negativeChooser.setDefaultOption("Positive", false);
-		negativeChooser.addOption("Negative", true);
-
-		SmartDashboard.putData("is Negative", negativeChooser);
-	}
-
-	public static AutoAlign getInstance() {
-		if (instance == null) {
-			instance = new AutoAlign();
-		}
-		return instance;
-	}
-
-	@Override
-	protected void runState() {
-		reefPose = DriverStation.getAlliance().isPresent() && DriverStation.getAlliance().get() == Alliance.Red ? new Pose2d(13.08, 4, new Rotation2d()) : new Pose2d(4.49, 4, new Rotation2d());
-		isRedAlliance = DriverStation.getAlliance().isPresent() && DriverStation.getAlliance().get() == Alliance.Red;
-		
-		if (getState() == AutoAlignStates.OFF) return;
-		
-		goalPose = getState().getTargetPose();
-		targetPose = getState().getTargetPose();
-		if (!readyForClose()) {
-			Translation2d temp = reefPose.getTranslation().minus(drive.getPose().getTranslation());
-			targetPose = new Pose2d(targetPose.getTranslation(), Rotation2d.fromRadians(Math.atan2(temp.getY(), temp.getX())));
-		}
-
-		// if there is no collision, it will go to braindead AA, or use repulsor to
-		// avoid the collision
-		if (!checkForReefCollision()) {
-			repulsorActivated = false;
-			braindeadAutoAlign();
-		} else {
-			repulsorActivated = true;
-			repulsor.setGoal(targetPose.getTranslation());
-			repulsorAutoAlign(drive.getPose(), repulsor.getCmd(drive.getPose(), drive.getRobotRelativeSpeeds(), MAX_SPEED.in(MetersPerSecond), USE_GOAL, targetPose.getRotation()));
-		}
-		logOutput();
-	}
-	
-	private void braindeadAutoAlign() {
-		Pose2d drivePose = drive.getPose();
-
-		if (!enteredBrainDead) {
-			translationXController.reset();
-		}
-
-		xApplied = translationXController.calculate(drivePose.getX(), targetPose.getX());
-		yApplied = translationYController.calculate(drivePose.getY(), targetPose.getY());
-
-		double rotationApplied = rotationController.calculate(drivePose.getRotation().getRadians(), targetPose.getRotation().getRadians());
+    @Override
+    protected void runState() {
+        if (getState() == AutoAlignStates.OFF) return;
 
 
+        Pose2d targetPose = targetPoseSupplier.get();
+        Pose2d currentPose = currentPoseSupplier.get();
 
+        Logger.recordOutput("AutoAlign/Target Pose", targetPose);
+        Logger.recordOutput("AutoAlign/Current Pose", currentPose);
 
-		if (isRedAlliance) drive.driveFieldRelative(-xApplied, -yApplied, rotationApplied, false, false); // was negative
-		else drive.driveFieldRelative(xApplied, yApplied, rotationApplied, false, false);
-	}
+        // Apply scalar drive around lalala
+        double currentDistance = currentPose.getTranslation().getDistance(targetPose.getTranslation());
+        double ffScaler = MathUtil.clamp(
+            (currentDistance-ffMinRadius) / (ffMaxRadius - ffMinRadius), 0.0, 1.0);
+        
+        driveErrorAbs = currentDistance;
+        translationalController.reset(lastSetpointTranslation.getDistance(targetPose.getTranslation()), translationalController.getSetpoint().velocity);
 
-	private void repulsorAutoAlign(Pose2d pose, SwerveSample sample) {
-		var targetSpeeds = sample.getChassisSpeeds();
-		targetSpeeds.vxMetersPerSecond += repulsionTranslationController.calculate(pose.getX(), sample.x);
-		targetSpeeds.vyMetersPerSecond += repulsionTranslationController.calculate(pose.getY(), sample.y);
-		targetSpeeds.omegaRadiansPerSecond += repulsionRotationController.calculate(pose.getRotation().getRadians(), sample.heading);
+        // drive error to 0 
+        double translationVelocityScalar = translationalController.getSetpoint().velocity * ffScaler
+            + translationalController.calculate(driveErrorAbs, 0.0);
 
-		Logger.recordOutput("TESTING VX", targetSpeeds.vxMetersPerSecond);
-		Logger.recordOutput("TESTING VY", targetSpeeds.vyMetersPerSecond);
-		if (isRedAlliance) drive.driveFieldRelative(-targetSpeeds.vxMetersPerSecond, -targetSpeeds.vyMetersPerSecond, targetSpeeds.omegaRadiansPerSecond, false, false); // was negative
-		else drive.driveFieldRelative(targetSpeeds.vxMetersPerSecond, targetSpeeds.vyMetersPerSecond, targetSpeeds.omegaRadiansPerSecond, false, false);
-	}
+        if (currentDistance < translationalController.getPositionTolerance()) {
+            translationVelocityScalar = 0;
+        }
 
-	private boolean checkForReefCollision() {
-		Pose2d currentPose = drive.getPose();
+        lastSetpointTranslation = new Pose2d(
+            targetPose.getTranslation(),
+            currentPose.getTranslation().minus(targetPose.getTranslation()).getAngle())
+                .transformBy(
+                        MathHelpers.transform2dFromTranslation(
+                                new Translation2d(translationalController.getSetpoint().position, 0.0)))
+                .getTranslation();
 
-		double t = calculateClosestPoint(currentPose, targetPose);
-		interpolatedPose = currentPose.interpolate(targetPose, t);
+        // Spin around
+        double thetaVelocity = rotationController.getSetpoint().velocity * ffScaler
+            + rotationController.calculate(
+                currentPose.getRotation().getRadians(),
+                targetPose.getRotation().getRadians());
 
-		interpolatedDistanceFromReef = interpolatedPose.getTranslation().getDistance(reefPose.getTranslation());
-		return interpolatedDistanceFromReef < REEF_HITBOX.in(Meters) + ROBOT_RADIUS.in(Meters);
-	}
+        thetaErrorAbs = Math.abs(currentPose.getRotation().minus(targetPose.getRotation()).getRadians());   
+        if (thetaErrorAbs < rotationController.getPositionTolerance()) {
+            thetaVelocity = 0;
+        }
+        
+        // Drive around
+         var translationVelocity = 
+                MathHelpers.pose2dFromRotation(currentPose.getTranslation().minus(targetPose.getTranslation()).getAngle())
+                .transformBy(MathHelpers.transform2dFromTranslation(new Translation2d(translationVelocityScalar, 0.0)))
+                .getTranslation();
+       
+        drive.driveFieldRelative(translationVelocity.getX(), translationVelocity.getY(), thetaVelocity, false, false);
+    
 
-	private double calculateClosestPoint(Pose2d startPose, Pose2d endPose) {
-		double numeratorX = (reefPose.getX() - startPose.getX()) * (endPose.getX() - startPose.getX());
-		double numeratorY = (reefPose.getY() - startPose.getY()) * (endPose.getY() - startPose.getY());
-		double denominator = Math.pow(endPose.getX() - startPose.getX(), 2) + Math.pow(endPose.getY() - startPose.getY(), 2);
+        
 
-		return MathUtil.clamp((numeratorX + numeratorY) / denominator, 0, 1);
-	}
-	
-	public boolean nearGoal() {
-		return drive.getPose().getTranslation().getDistance(goalPose.getTranslation()) < DISTANCE_ERROR_MARGIN.in(Meters) && (Math.abs(repulsionRotationController.getError()) < ANGLE_ERROR_MARGIN.in(Radians) || Math.abs(rotationController.getPositionError()) < ANGLE_ERROR_MARGIN.in(Radians));
-	}
-
-	public boolean nearGoalSource() {
-		return autoAlignDebouncer.calculate(
-			drive.getPose().getTranslation().getDistance(goalPose.getTranslation()) < DISTANCE_ERROR_MARGIN.in(Meter) && (Math.abs(repulsionRotationController.getError()) < ANGLE_ERROR_MARGIN.in(Radians) || Math.abs(rotationController.getPositionError()) < ANGLE_ERROR_MARGIN.in(Radians))
-			);
-	}
-	
-	public boolean timedOut() {
-		if (getStateTime() > 0.01 && drive.getPose().getTranslation().getDistance(targetPose.getTranslation()) < MOVEMENT_THRESHOLD) {
-			timer = getStateTime();
-		}
-		
-		if (timer != -1 && getStateTime() - timer > TIMEOUT_THRESHOLD) {
-			timer = -1;
-			return true;
-		} else return false;
-	}
-	
-	public void resetPID() { //TODO call this every time entering AA far
-		Pose2d currentPose = drive.getPose();
-		ChassisSpeeds currentSpeed = ChassisSpeeds.fromRobotRelativeSpeeds(drive.getRobotRelativeSpeeds(), currentPose.getRotation());
-
-		translationXController.reset();
-		translationYController.reset();
-		rotationController.reset(currentSpeed.omegaRadiansPerSecond);
-		repulsionTranslationController.reset();
-		repulsionRotationController.reset();
-
-		enteredBrainDead = false;
-	}
-	
-	public boolean readyForClose() {
-		return (drive.getPose().getTranslation().getDistance(goalPose.getTranslation()) < getState().getDistanceForCloseAA().in(Meters));
-	}
-
-	private void logOutput() {
-		Pose2d[] arrowsArray = new Pose2d[] {};
-	
-		Logger.recordOutput("AutoAlign/State", getState().getStateString());
-		Logger.recordOutput("AutoAlign/Target Pose", targetPose);
-		Logger.recordOutput("AutoAlign/Interpolated Pose", interpolatedPose);
-		Logger.recordOutput("AutoAlign/Interpolated Distance From Reef", interpolatedDistanceFromReef);
-		Logger.recordOutput("AutoAlign/ReefPose", reefPose);
-		Logger.recordOutput("AutoAlign/Repulsor Activated", repulsorActivated);
-		Logger.recordOutput("AutoAlign/Arrows", repulsor.getArrows().toArray(arrowsArray));
-	}
+        
+    }
 }
